19891 - reduction common added.
