+++ ttnn/cpp/ttnn/operations/reduction/topk/topk.cpp
@@ -15,18 +15,12 @@
 #include "ttnn/operations/data_movement/transpose/transpose.hpp"
 #include "ttnn/operations/data_movement/common/common.hpp"
 #include "ttnn/operations/data_movement/fill_pad/fill_pad.hpp"
+#include "ttnn/operations/reduction/reduction_common/reduction_common.hpp"
 
 namespace ttnn::operations::reduction {
 
 namespace {
 namespace CMAKE_UNIQUE_NAMESPACE {
-
-template <class Tuple, class T = std::decay_t<std::tuple_element_t<0, std::decay_t<Tuple>>>>
-std::vector<std::optional<T>> tuple_to_vector_optional(Tuple&& tuple) {
-    return std::apply(
-        [](auto&&... elems) { return std::vector<std::optional<T>>{std::forward<decltype(elems)>(elems)...}; },
-        std::forward<Tuple>(tuple));
-}
 
 uint32_t get_nearest_supported_k_value(uint32_t k) {
     // LLK only support k = 32, 64 for now
@@ -35,15 +29,6 @@ uint32_t get_nearest_supported_k_value(uint32_t k) {
     } else {
         return 64;
     }
-}
-
-Tensor perform_transpose(
-    const Tensor& input_tensor, const bool is_dim_last_idx, const int8_t dim1 = -1, const int8_t dim2 = -1) {
-    return is_dim_last_idx ? input_tensor : ttnn::transpose(input_tensor, dim1, dim2, input_tensor.memory_config());
-}
-
-Tensor transform_to_4d_tensor(const Tensor& input_tensor, const bool is_rank_le_4d) {
-    return is_rank_le_4d ? ttnn::unsqueeze_to_4D(input_tensor) : data_movement::squeeze_from_ND_to_4D(input_tensor);
 }
 
 // one stop for all transformations needed after executing top-k
@@ -120,9 +105,9 @@ std::vector<Tensor> ExecuteTopK::invoke(
     // K must be a supported shape
     uint32_t adjusted_k = CMAKE_UNIQUE_NAMESPACE::get_nearest_supported_k_value(k);
     // if dim is not last dimension, transpose it
-    Tensor transposed_tensor = CMAKE_UNIQUE_NAMESPACE::perform_transpose(input_tensor, is_dim_last_idx, dim, -1);
+    Tensor transposed_tensor = reduction_common::perform_transpose(input_tensor, is_dim_last_idx, dim, -1);
     // if input is not 4d, convert it to 4d
-    Tensor transformed_tensor = CMAKE_UNIQUE_NAMESPACE::transform_to_4d_tensor(transposed_tensor, is_rank_le_4d);
+    Tensor transformed_tensor = reduction_common::transform_to_4d_tensor(transposed_tensor, is_rank_le_4d);
     // add padding if needed
     Tensor padded_tensor = ttnn::fill_implicit_tile_padding(
         transformed_tensor, largest ? std::numeric_limits<float>::min() : std::numeric_limits<float>::max());
@@ -132,7 +117,7 @@ std::vector<Tensor> ExecuteTopK::invoke(
         {padded_tensor},
         {},
         optional_output_tensors.has_value()
-            ? CMAKE_UNIQUE_NAMESPACE::tuple_to_vector_optional(optional_output_tensors.value())
+            ? reduction_common::tuple_to_vector_optional(optional_output_tensors.value())
             : std::vector<std::optional<Tensor>>{},
         queue_id);
